requirements.txt
# ConductAI Documentation Pipeline - Dependencies
# All required packages for the complete system

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# Manim (target library)
manim>=0.17.3

# AI/LLM Dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0
anthropic>=0.18.0
openai>=1.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Data Processing & Utilities
pydantic>=2.0.0
PyYAML>=6.0
json5>=0.9.0
pathlib
asyncio

# Web Scraping & Document Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyPDF2>=3.0.0

# Scientific Computing & Embeddings
numpy>=1.24.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Console & CLI Enhancement
rich>=13.0.0
typer>=0.9.0

# Async Operations
aiofiles>=23.0.0

# Development & Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Note: MCP Desktop Commander is installed via npx, not pip
# Run: npx @modelcontextprotocol/desktop-commander-server 